Refactor for MVC
Move take_action to cmd.py to allow it to accessing stdout and stderr. Allow tests to directly access managers.
