+++ python/ssrcli/managers.py
@@ -7,7 +7,7 @@ import requests
 
 from .config import config
 from .models import SsrConf, SsrSub
-from .exceptions import SsrcliException, SsrUrlInvalid, RequireMoreParam, InvalidParam, NoSuchOperation
+from .exceptions import SsrcliException, SsrUrlInvalid
 
 Models = Union[SsrConf, SsrSub]
 Param = Union[str, List[int]]
@@ -172,73 +172,3 @@ class SsrSubManager(Manager):
         results = asyncio.get_event_loop().run_until_complete(asyncio.gather(*tasks))
         for result in results:
             self._update_result(result, ins_dict[result[1]])
-
-
-def take_action(param_dict: Dict[str, Param]) -> None:
-    try:
-        model, manager = {
-            'conf': (SsrConf, SsrConfManager()),
-            'sub': (SsrSub, SsrSubManager()),
-        }[param_dict['model']]
-
-        action = param_dict['action']
-    except KeyError as error:
-        raise RequireMoreParam(error.args[0])
-
-    try:
-        if action in ['ls', 'get']:  # `get` is deprecated
-            if action == 'get':
-                print('`get` is deprecated! Use `ls` as it can work the same as `get`')
-            id_list = param_dict.get('ins_id', None) if param_dict['all'] else None
-            for ins in manager.list(id_list):
-                print(ins)
-            if param_dict['current']:
-                print('\nCurrently-used configuration is:')
-                print(manager.load_use())
-
-        elif action == 'add':
-            if model == SsrConf:
-                if 'ssr_url' in param_dict.keys():
-                    json_info = from_ssr_url(param_dict['ssr_url'])
-                else:
-                    try:
-                        json_info = json.loads(param_dict['ins_json'])
-                    except ValueError:
-                        raise InvalidParam('ins_json')
-                print(manager.create(**json_info))
-
-            elif model == SsrSub:
-                try:
-                    json_info = json.loads(param_dict['ins_json'])
-                except ValueError:
-                    raise InvalidParam('ins_json')
-                print(manager.create(**json_info))
-
-            else:
-                raise RuntimeError()
-
-        elif action == 'rm':
-            if param_dict['all']:
-                id_list = None
-            else:
-                id_list = param_dict['ins_id']
-            manager.delete(id_list)
-
-        elif action == 'edit':
-            raise NotImplementedError()  # TODO(myl7)
-
-        elif action == 'take':
-            if model != SsrConf:
-                raise NoSuchOperation('take')
-            manager.take_use(param_dict['ins_id'].pop())
-
-        elif action == 'update':
-            if model != SsrSub:
-                raise NoSuchOperation('update')
-            id_list = param_dict.get('ins_id', None) if param_dict['all'] else None
-            manager.update(id_list)
-
-        else:
-            raise RuntimeError()
-    except KeyError as error:
-        raise RequireMoreParam(error.args[0])
